revert change of serialized type
using the Tile interface{} instead of HexCoord for the player allows us to omitempty when it is nil/zero, however it adds a layer of indirection which alters how reflect.DeepEqual compares the values...

I could change all the test cases over but perhaps it's better to avoid omitting the player's initial position, even when it's (0, 0).

While in here, changed the struct's property name (was Player) to match the name I've chosen for the init-value in the JSON representation (ichiban).  I made both of these up, my only preference is to avoid using the term "man" unnecessarily.