authored my first hexaban puzzle!
the design is inspired by level 25 from the authoritative collection of 90 xsokoban levels, the original is solvable by Rolling Stone but is the third most challenging even with all of the search enhancements on.  I have attempted to retain the particular features (corners, clusters and tunnels) that make the original challenging while packing it into the features which make hexagonal variants interesting.  I hope I've succeeded!

This will make an interesting test bed for my hex-ban solver, but it also looks quite pleasing when rendered properly (not with this text format).