rename hexaban -> hexoban everywhere
 - search/replace: s/([Hh])exaban/$1exoban/g
 - Verified that existing tests pass.

I thought it would be cute to take the Greek prefix instead of the 'o' but that was before discovering that DWS had already invented the puzzle and called it hexoban.  Then I continued with the 'a' form thinking it would offer some uniqueness, but it turns out there's an Indian company with a pesticide of the same name, and there's no overlap but I don't want to fight for ad placement with it either.

Anyway, adopting the 'o' variant offers some consistency with existing puzzle creations and avoids the misunderstanding that I might not be aware of the earlier invention by DWS and contributions of others (which, of course, many README's will contradict, but who is reading those?)