Finish defining the puzzle editor/creator by handling CLI inputs for attribute values and the puzzle map definition
Entering the map via CLI has problems on Windows, I think because of how PowerShell forwards the CLI input, but fixing is a low priority... it should still work on Linux (the parsing and CLI logic hasn't changed there, only a thin wrapper for it was added to the Builder type.  Looks like Viper and other CLI utility libraries also have a helluva time with Windows terminals.  bugger.